pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quantlix"
version = "0.1.0"
description = "Quantlix — AI-focused GPU inference platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.9.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "rich>=13.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "minio>=7.2.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.1.0",
    "prometheus-client>=0.19.0",
    "aiosmtplib>=3.0.0",
    "stripe>=8.0.0",
]

[project.scripts]
quantlix = "cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "orchestrator*", "cli*", "sdk*"]
